# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Globapay Platform API
  description: Multi-tenant payments orchestration API with contract-first design
  version: 0.1.0
  contact:
    name: Globapay Platform Team
    url: https://globapay.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.globapay.com
    description: Production server
security:
  - bearerAuth: []
  - apiKey: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Unique key for request idempotency
      required: false
      schema:
        type: string
        maxLength: 255
    Cursor:
      name: cursor
      in: query
      description: Cursor for pagination
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return (1-100)
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    ValidationError:
      description: Validation Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: Problem type identifier
        title:
          type: string
          description: Human-readable problem summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable problem details
        instance:
          type: string
          format: uri
          description: Problem instance identifier
      required: [type, title, status]
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true
      required: [data, hasMore, nextCursor]

    # Entity Schemas
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchantId:
          type: string
          format: uuid
        paymentLinkId:
          type: string
          format: uuid
          nullable: true
        checkoutSessionId:
          type: string
          format: uuid
          nullable: true
        reference:
          type: string
          maxLength: 255
          nullable: true
        amount:
          type: integer
          minimum: 1
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        status:
          type: string
          enum: [pending, authorized, captured, failed, cancelled, refunded, partially_refunded]
        paymentMethod:
          type: object
          properties:
            type:
              type: string
              enum: [card, bank_transfer, wallet]
            details:
              type: object
              properties:
                last4:
                  type: string
                  pattern: '^[0-9]{4}$'
                brand:
                  type: string
                country:
                  type: string
                  pattern: '^[A-Z]{2}$'
              additionalProperties: false
          required: [type]
          additionalProperties: false
        customer:
          type: object
          properties:
            email:
              type: string
              format: email
            name:
              type: string
              maxLength: 255
            phone:
              type: string
              maxLength: 20
          additionalProperties: false
        fraudScore:
          type: number
          minimum: 0
          maximum: 100
          nullable: true
        threeDsResult:
          type: object
          properties:
            status:
              type: string
              enum: [authenticated, not_authenticated, attempted, unavailable, rejected]
            version:
              type: string
              enum: ['1.0', '2.0', '2.1', '2.2']
          required: [status]
          additionalProperties: false
          nullable: true
        fees:
          type: object
          properties:
            processing:
              type: integer
              minimum: 0
            platform:
              type: integer
              minimum: 0
          additionalProperties: false
          nullable: true
        pspTransactionId:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
        processedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, merchantId, amount, currency, status, paymentMethod, createdAt, updatedAt]
    RefundRequest:
      type: object
      properties:
        amount:
          type: integer
          minimum: 1
          description: Amount to refund in smallest currency unit
        reason:
          type: string
          enum: [duplicate, fraudulent, requested_by_customer, other]
        description:
          type: string
          maxLength: 500
        metadata:
          type: object
          additionalProperties:
            type: string
      required: [amount, reason]
    Refund:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        merchantId:
          type: string
          format: uuid
        amount:
          type: integer
          minimum: 1
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        reason:
          type: string
          enum: [duplicate, fraudulent, requested_by_customer, other]
        description:
          type: string
          maxLength: 500
          nullable: true
        status:
          type: string
          enum: [pending, succeeded, failed, cancelled]
        pspRefundId:
          type: string
          nullable: true
        initiatedBy:
          type: object
          properties:
            userId:
              type: string
              format: uuid
            type:
              type: string
              enum: [manual, automatic, chargeback]
          required: [type]
          additionalProperties: false
        metadata:
          type: object
          additionalProperties:
            type: string
        processedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, transactionId, merchantId, amount, currency, reason, status, initiatedBy, createdAt, updatedAt]
...
paths:
...
  /transactions:
    get:
      summary: List transactions
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          description: Filter by transaction status
          schema:
            type: string
            enum: [pending, authorized, captured, failed, cancelled, refunded, partially_refunded]
        - name: merchantId
          in: query
          description: Filter by merchant ID
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          description: Filter transactions from this date
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: Filter transactions to this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /transactions/{id}:
    get:
      summary: Get transaction details
      operationId: getTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /transactions/{id}/refund:
    post:
      summary: Refund a transaction
      operationId: refundTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '201':
          description: Refund initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  # Incoming Webhooks
...